requirements.txt
# ComfyUI Substance Designer Integration Plugin Requirements

# Core dependencies
torch>=1.9.0
numpy>=1.20.0
Pillow>=8.0.0

# Optional dependencies for enhanced functionality
# These are typically already available in ComfyUI environments

# For JSON handling (built-in to Python)
# For subprocess management (built-in to Python)
# For file operations (built-in to Python)
# For threading and concurrent processing (built-in to Python)

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0

